Refactored layout a bit
