+++ package.json
@@ -17,6 +17,12 @@
   },
   "devDependencies": {
     "bulma": "^0.8.0",
+    "eslint": "^6.8.0",
+    "eslint-config-airbnb": "^18.0.1",
+    "eslint-plugin-import": "^2.19.1",
+    "eslint-plugin-jsx-a11y": "^6.2.3",
+    "eslint-plugin-react": "^7.17.0",
+    "eslint-plugin-react-hooks": "^1.7.0",
     "node-sass": "^4.9.3"
   },
   "browserslist": {
